Phased AuditRunner execution

+++ src/main/java/org/libj/math/BigIntValue.java
@@ -51,34 +51,39 @@ abstract class BigIntValue extends Number {
     }
     else {
       final boolean useCritical = ManagementFactory.getRuntimeMXBean().getInputArguments().toString().indexOf("-Xcomp") > 0;
-      NATIVE_THRESHOLD = useCritical ? 0 : 15;
 
       final String fileName = "libmath" + (useCritical ? "c" : "j");
       final String extension = ".so"; // FIXME: Detect appropriate platform here
 
       final URL url = BigIntValue.class.getResource("/" + fileName + extension);
-      final File file;
-      try {
-        if (url.toString().startsWith("jar:file:")) {
-          final Path tempPath = Files.createTempFile(fileName, extension);
-          file = tempPath.toFile();
-          file.deleteOnExit();
-          try (final InputStream in = url.openStream()) {
-            Files.copy(in, tempPath, StandardCopyOption.REPLACE_EXISTING);
+      if (url == null) {
+        NATIVE_THRESHOLD = Integer.MAX_VALUE;
+      }
+      else {
+        NATIVE_THRESHOLD = useCritical ? 0 : 15;
+        final File file;
+        try {
+          if (url.toString().startsWith("jar:file:")) {
+            final Path tempPath = Files.createTempFile(fileName, extension);
+            file = tempPath.toFile();
+            file.deleteOnExit();
+            try (final InputStream in = url.openStream()) {
+              Files.copy(in, tempPath, StandardCopyOption.REPLACE_EXISTING);
+            }
+          }
+          else if (url.toString().startsWith("file:")) {
+            file = new File(url.getPath());
+          }
+          else {
+            throw new ExceptionInInitializerError("Unsupported protocol: " + url);
           }
         }
-        else if (url.toString().startsWith("file:")) {
-          file = new File(url.getPath());
+        catch (final IOException e) {
+          throw new ExceptionInInitializerError(e);
         }
-        else {
-          throw new ExceptionInInitializerError("Unsupported protocol: " + url);
-        }
-      }
-      catch (final IOException e) {
-        throw new ExceptionInInitializerError(e);
-      }
 
-      System.load(file.getAbsolutePath());
+        System.load(file.getAbsolutePath());
+      }
     }
   }
